# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "identity-verificate-service",
        "version": "1.0",
        "description": "本人性確認サービス"
    },
    "servers": [
        {
            "url": "/identity-verificate"
        }
    ],
    "paths": {
...
        "/collate": {
            "post": {
                "tags": [
                    "本人性確認コード照合"
                ],
                "description": "本人性確認コード照合を行い、照合結果を取得します。",
                "requestBody": {
                    "description": "照合結果",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "照合する本人性確認コード",
                                "properties": {
                                    "identifyCode": {
                                        "type": "string",
                                        "description": "本人性確認コード"
                                    }
                                },
                                "example": {
                                    "identifyCode": "ukO8z+Xf8vv7yxXQj2Hpo"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "照合に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "照合結果",
                                    "properties": {
                                        "pxrId": {
                                            "type": "string",
                                            "description": "PXR-ID"
                                        },
                                        "userId": {
                                            "type": "string",
                                            "description": "利用者ID"
                                        },
                                        "actor": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "アクターカタログコードオブジェクト"
                                        },
                                        "wf": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "WF定義カタログコードオブジェクト"
                                        },
                                        "app": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "APP定義カタログコードオブジェクト"
                                        }
                                    },
                                    "example": {
                                        "pxrId": "sample.pxrid",
                                        "userId": "123456",
                                        "actor": {
                                            "_value": 1000003,
                                            "_ver": 1
                                        },
                                        "wf": {
                                            "_value": 1000004,
                                            "_ver": 1
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ind/collate": {
            "post": {
                "tags": [
                    "個人による本人性確認コード照合"
                ],
                "description": "本人性確認コード照合を行い、照合結果を取得します。",
                "requestBody": {
                    "description": "照合結果",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "照合する本人性確認コード",
                                "properties": {
                                    "identifyCode": {
                                        "type": "string",
                                        "description": "本人性確認コード"
                                    }
                                },
                                "example": {
                                    "identifyCode": "ukO8z+Xf8vv7yxXQj2Hpo"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "照合に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "照合結果",
                                    "properties": {
                                        "pxrId": {
                                            "type": "string",
                                            "description": "PXR-ID"
                                        },
                                        "userId": {
                                            "type": "string",
                                            "description": "利用者ID"
                                        },
                                        "actor": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "アクターカタログコードオブジェクト"
                                        },
                                        "wf": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "WF定義カタログコードオブジェクト"
                                        },
                                        "app": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "APP定義カタログコードオブジェクト"
                                        }
                                    },
                                    "example": {
                                        "pxrId": "sample.pxrid",
                                        "userId": "123456",
                                        "actor": {
                                            "_value": 1000003,
                                            "_ver": 1
                                        },
                                        "wf": {
                                            "_value": 1000004,
                                            "_ver": 1
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}